+++ packages/contracts/src/libraries/LibRelationship.sol
@@ -23,8 +23,8 @@ library LibRelationship {
     IWorld world,
     IUintComp components,
     uint256 accountID,
-    uint256 npcIndex,
-    uint256 relIndex
+    uint32 npcIndex,
+    uint32 relIndex
   ) internal returns (uint256) {
     uint256 id = world.getUniqueEntityId();
     IsRelationshipComponent(getAddressById(components, IsRelCompID)).set(id);
@@ -44,8 +44,8 @@ library LibRelationship {
   function canCreate(
     IUintComp components,
     uint256 accountID,
-    uint256 npcIndex,
-    uint256 relIndex
+    uint32 npcIndex,
+    uint32 relIndex
   ) internal view returns (bool) {
     uint256 registryID = LibRegistryRelationship.get(components, npcIndex, relIndex);
     if (isBlacklisted(components, accountID, registryID)) return false;
@@ -57,8 +57,8 @@ library LibRelationship {
   function has(
     IUintComp components,
     uint256 accountID,
-    uint256 npcIndex,
-    uint256 relIndex
+    uint32 npcIndex,
+    uint32 relIndex
   ) internal view returns (bool) {
     return get(components, accountID, npcIndex, relIndex) != 0;
   }
@@ -69,8 +69,8 @@ library LibRelationship {
     uint256 accountID,
     uint256 registryID
   ) internal view returns (bool) {
-    uint256[] memory blacklist = LibRegistryRelationship.getBlacklist(components, registryID);
-    uint256 npcIndex = LibRegistryRelationship.getNpcIndex(components, registryID);
+    uint32[] memory blacklist = LibRegistryRelationship.getBlacklist(components, registryID);
+    uint32 npcIndex = LibRegistryRelationship.getNpcIndex(components, registryID);
     for (uint256 i = 0; i < blacklist.length; i++) {
       if (has(components, accountID, npcIndex, blacklist[i])) return true;
     }
@@ -84,8 +84,8 @@ library LibRelationship {
     uint256 accountID,
     uint256 registryID
   ) internal view returns (bool) {
-    uint256[] memory whitelist = LibRegistryRelationship.getWhitelist(components, registryID);
-    uint256 npcIndex = LibRegistryRelationship.getNpcIndex(components, registryID);
+    uint32[] memory whitelist = LibRegistryRelationship.getWhitelist(components, registryID);
+    uint32 npcIndex = LibRegistryRelationship.getNpcIndex(components, registryID);
     if (whitelist.length == 0) return true;
     for (uint256 i = 0; i < whitelist.length; i++) {
       if (has(components, accountID, npcIndex, whitelist[i])) return true;
@@ -99,8 +99,8 @@ library LibRelationship {
   function get(
     IUintComp components,
     uint256 accountID,
-    uint256 npcIndex,
-    uint256 relIndex
+    uint32 npcIndex,
+    uint32 relIndex
   ) internal view returns (uint256 result) {
     QueryFragment[] memory fragments = new QueryFragment[](4);
     fragments[0] = QueryFragment(QueryType.Has, getComponentById(components, IsRelCompID), "");
