+++ packages/contracts/src/systems/AccountUpgradeSkillSystem.sol
@@ -18,7 +18,7 @@ contract AccountUpgradeSkillSystem is System {
   constructor(IWorld _world, address _components) System(_world, _components) {}
 
   function execute(bytes memory arguments) public returns (bytes memory) {
-    uint256 skillIndex = abi.decode(arguments, (uint256));
+    uint32 skillIndex = abi.decode(arguments, (uint32));
     uint256 accountID = LibAccount.getByOperator(components, msg.sender);
     uint256 registryID = LibRegistrySkill.getByIndex(components, skillIndex);
 
@@ -50,7 +50,7 @@ contract AccountUpgradeSkillSystem is System {
     return "";
   }
 
-  function executeTyped(uint256 skillIndex) public returns (bytes memory) {
+  function executeTyped(uint32 skillIndex) public returns (bytes memory) {
     return execute(abi.encode(skillIndex));
   }
 }
